requirements.txt
Flask
Flask-SocketIO
pandas
smtplib
openai
sendgrid
celery
redis
google-api-python-client
Flask-Migrate
